Fix: change dashboard to scrap
### [Fix]: [change dashboard to scrap]

# 1. 대시보드 화면을 scrap화면으로 사용하기 위하여 Rename
