+++ src/BlogApp/app/src/main/java/com/smilegateblog/smliegateblog/presentation/main/scrap/ScrapFragment.kt
@@ -1,0 +1,36 @@
+package com.smilegateblog.smliegateblog.presentation.main.scrap
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.smilegateblog.smliegateblog.databinding.FragmentScrapBinding
+
+
+class ScrapFragment : Fragment() {
+
+    private var _binding: FragmentScrapBinding? = null
+
+    // This property is only valid between onCreateView and
+    // onDestroyView.
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+
+        _binding = FragmentScrapBinding.inflate(inflater, container, false)
+        val root: View = binding.root
+
+
+        return root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
