+++ src/BlogApp/app/src/main/java/com/smilegateblog/smliegateblog/presentation/main/scrap/ScrapViewModel.kt
@@ -1,4 +1,4 @@
-package com.smilegateblog.smliegateblog.presentation.main.dashboard
+package com.smilegateblog.smliegateblog.presentation.main.scrap
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -8,10 +8,10 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class DashboardViewModel @Inject constructor(private val loginUseCase: LoginUseCase) : ViewModel() {
+class ScrapViewModel @Inject constructor(private val loginUseCase: LoginUseCase) : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is dashboard Fragment"
+        value = "This is scrap Fragment"
     }
     val text: LiveData<String> = _text
 }
