+++ src/BlogApp/app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <resources>
     <string name="app_name">SmlieGateBlogApp</string>
     <string name="title_home">Home</string>
-    <string name="title_dashboard">Dashboard</string>
+    <string name="title_scrap">scrap</string>
     <string name="title_notifications">Notifications</string>
